refactor(memory): move extract.rs inline tests to sibling extract_tests.rs
The develop merge brought the frozen file-budget and inline-test baselines
(#1974); extract.rs stood at 1869 lines with 234 inline test-module lines
against frozen 1738/182. Moving the extractor-http test module to the
sibling-file convention brings extract.rs to 1638 lines and zero inline
test debt; both baselines are ratcheted down accordingly.
